factoring user profile endpoints out of UserController
